Refactor RegisterPage
